CARDS-2656: Refactor login related components to functional Refactored loginDialogue to loginDialog
