+++ modules/data-entry/src/main/frontend/src/dataHomepage/LiveTable.jsx
@@ -25,7 +25,7 @@ import { Link } from 'react-router';
 import { DateTime } from "luxon";
 import Filters from "./Filters.jsx";
 import { getEntityIdentifier } from "../themePage/EntityIdentifier.jsx";
-import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialog.js";
 
 import LiveTableStyle from "./tableStyle.jsx";
