+++ modules/data-entry/src/main/frontend/src/questionnaire/SubjectTimeline.jsx
@@ -37,7 +37,7 @@ import {
 } from "@mui/lab";
 
 import DateQuestionUtilities from "./DateQuestionUtilities.jsx";
-import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialog.js";
 import QuestionnaireStyle from "./QuestionnaireStyle.jsx";
 import { displayQuestion } from "./Subject.jsx";
 import { ENTRY_TYPES, QUESTION_TYPES, SECTION_TYPES } from "./FormEntry.jsx"
