+++ modules/homepage/src/main/frontend/src/adminDashboard/AdminResourceListing.jsx
@@ -20,7 +20,7 @@ import React, { useState, useContext, useEffect } from "react";
 
 import AdminScreen from "./AdminScreen.jsx";
 import NewItemButton from "../components/NewItemButton.jsx";
-import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialog.js";
 import MaterialReactTable from "material-react-table";
