+++ modules/patient-portal/src/main/frontend/src/patient-portal/Clinics.jsx
@@ -29,7 +29,7 @@ import {
 import Fields from "../questionnaireEditor/Fields.jsx";
 import AdminResourceListing from "../adminDashboard/AdminResourceListing.jsx";
 import ResponsiveDialog from "../components/ResponsiveDialog.jsx";
-import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialog.js";
 import { camelCaseToWords } from "../questionnaireEditor/LabeledField.jsx";
 
 function Clinics(props) {
