+++ modules/statistics/src/main/frontend/src/Statistics/UserStatistics.jsx
@@ -23,7 +23,7 @@ import {
   Typography
 } from "@mui/material";
 import { withStyles } from 'tss-react/mui';
-import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "../login/loginDialog.js";
 import statisticsStyle from "./statisticsStyle.jsx";
 import Statistic from "./Statistic.jsx";
