+++ modules/variants/src/main/frontend/src/variantFilesContainer.jsx
@@ -44,7 +44,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { DateTime } from "luxon";
 import DragAndDrop from "./components/dragAndDrop.jsx";
 import { escapeJQL } from "./escape.jsx";
-import { fetchWithReLogin, GlobalLoginContext } from "./login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "./login/loginDialog.js";
 
 const useStyles = makeStyles()(theme => ({
   root: {
