+++ modules/vocabularies/src/main/frontend/src/owlInstaller.jsx
@@ -30,7 +30,7 @@ import {
 
 import { makeStyles } from 'tss-react/mui';
 
-import { fetchWithReLogin, GlobalLoginContext } from "./login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "./login/loginDialog.js";
 
 const useStyles = makeStyles()(theme => ({
   buttonProgress: {
