+++ modules/vocabularies/src/main/frontend/src/vocabularyAction.jsx
@@ -33,7 +33,7 @@ import {
 
 import { makeStyles } from 'tss-react/mui';
 
-import { fetchWithReLogin, GlobalLoginContext } from "./login/loginDialogue.js";
+import { fetchWithReLogin, GlobalLoginContext } from "./login/loginDialog.js";
 
 const Phase = require("./phaseCodes.json");
